backend/requirements.txt
Django>=4.2
djangorestframework
requests
psycopg2-binary
pytest
pytest-django
django-cors-headers
